// codigo-fonte/frontend/src/api/models/list-user-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IgescConecta.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Filter } from './filter';

/**
 * 
 * @export
 * @interface ListUserRequest
 */
export interface ListUserRequest {
    /**
     * 
     * @type {number}
     * @memberof ListUserRequest
     */
    'pageNumber'?: number;
    /**
     * 
     * @type {number}
     * @memberof ListUserRequest
     */
    'pageSize'?: number;
    /**
     * 
     * @type {Array<Filter>}
     * @memberof ListUserRequest
     */
    'filters'?: Array<Filter>;
}